Added model classes that will be used for quests
Stage Model is a list of adventure cards that will represent the cards that each player plays for each tage of a quest. Can use the same base for the sponsor

Setup Model is a list of stages, and will hold all the cards played for each seperate stage. Again sponsor works similarly to how the players do.

Stage = list of adventure cards
Setup = list of stages

so all together
setup = list of list of cards